some changes, add stats, stats modifiers, new passives, board load

+++ Assets/Scripts/Gameplay/Combat/Contexts/BeforeHitContext.cs
@@ -6,7 +6,7 @@ namespace Project.Gameplay.Gameplay.Combat.Contexts
     {
         public Figure Attacker { get; set; }
         public Figure Target { get; set; }
-        public int BaseDamage { get; set; }
+        public float BaseDamage { get; set; }
         public float DamageMultiplier { get; set; } = 1f;
         public int BonusDamage { get; set; }
         public bool IsCritical { get; set; }
